Put all change in xrt_module.cpp. Not ideal...
